README.md
This is going to end up being a class that in general helps with a large amount of tedious calculations that would take time to do on the calculator that I would rather do in bulk all at once using a program. Feel free to use and expand as you choose.